requirements.txt
# API and Web framework
fastapi>=0.115.0
uvicorn>=0.26.0
starlette>=0.26.1
sse-starlette>=1.6.1

# HTTP client
httpx>=0.24.0

# Data processing and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
jsonschema>=4.17.3

# Tools and utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# MCP
mcp>=1.8.1

# n8n SDK
n8n-sdk-python>=0.1.2
